Feature notificacoes avaliacoes

+++ app/src/main/java/com/stack/open_work_mobile/services/OnAvaliarClickListener.kt
@@ -1,5 +1,5 @@
 package com.stack.open_work_mobile.Interface
 
 interface OnAvaliarClickListener {
-    fun onAvaliarClick(id: Long, nome: String, grade: Int)
+    fun onAvaliarClick(id: Int, nome: String, grade: Int)
 }
